+++ src/dotnet-interactive-vscode/insiders/.eslintignore
@@ -1,3 +1,3 @@
 jupyter.d.ts
-vscode.proposed.d.ts
+vscode.proposed.*.d.ts
 vscode.d.ts
